+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3
+
+* Change font to Albert Sans
+* Add `scale`, `opacity` and `colored` properties to `NoiseOverlay` widget
+
+## 0.0.2
+
+* Readme and workflow updates
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial unlisted release